refactor: Move database functions according to new structure
