+++ database/image.go
@@ -32,3 +32,8 @@ func DeleteImage(fileName string) (string, bool) {
 		return "", false
 	}
 }
+
+func RenameImage(oldFileName, newFileName string) error {
+	image := models.Image{}
+	return DB.Model(&image).Where("file_name = ?", oldFileName).Update("file_name", newFileName).Error
+}
